Resolves #25: Prevent repeated action executions
