+++ lib/vagrant-goodhosts/Action/UpdateHosts.rb
@@ -1,24 +1,11 @@
-require_relative "../GoodHosts"
 module VagrantPlugins
   module GoodHosts
     module Action
-      class UpdateHosts
-        include GoodHosts
-        @@updated = false
+      class UpdateHosts < BaseAction
 
-        def initialize(app, env)
-          @app = app
-          @machine = env[:machine]
-          @ui = env[:ui]
-        end
-
-        def call(env)
-          unless @@updated
-            @@updated = true
-            @ui.info "[vagrant-goodhosts] Checking for host entries"
-            addHostEntries()
-          end
-          @app.call(env)
+        def run(env)
+          @ui.info "[vagrant-goodhosts] Checking for host entries"
+          addHostEntries()
         end
 
       end
